Restructured server for getting shift information, added seperate file for shift on frontend
