+++ frontend/src/api/classesPageService.js
@@ -8,20 +8,4 @@ const getAllClasses = () => api.get('/classes')
         console.error(error);
     });
 
-const getShiftInfo = async (volunteerID, scheduleID, shiftDate) => {
-    try {
-        const response = await api.post('/classes/shifts', { 
-            volunteerID, 
-            scheduleID, 
-            shiftDate 
-        });
-
-        const res = response.data;
-        return res;  
-    } catch (error) {
-        console.error('Error fetching shift info:', error);
-        throw error;  
-    }
-}
-
-export { getAllClasses, getShiftInfo };
+export { getAllClasses };
